refactor(db): split inbox repository by owning table
crates/persistence/db/src/repositories/inbox.rs was 3751 lines (1987
production + a 1743-line test module). Split into a module directory
partitioned by the table each function owns:

- items: inbox_items CRUD, single-type sub-items, source-group membership
- source_groups: inbox_source_groups scan records
- classification: inbox_classifications, _evidence, inbox_file_overrides
- metadata: inbox_file_metadata, inbox_classification_breakdown
- plan_links: inbox_plan_links
- projections: read-only cross-table stats, list, and grouping-key queries

Behaviour-preserving move. mod.rs re-exports every public item, so all
repositories::inbox::* consumer paths are unchanged and no caller was
touched. The exclude_split_placeholder! query fragment moves to
projections, where its three in-crate callers live; it stays re-exported
crate-internally so q_desktop is untouched.

Test logic is unmodified: tests.rs is the old test module dedented, with
one added 'use sqlx::SqlitePool;' that the old 'use super::*' glob
previously supplied. Diff is 96.9% content-identical relocation.

Closes #968
